purchases export (#2047)
